refactor(routing): Moving to fast endpoint
Moving routing to fast endpoints lib

refactor(routing): Moving to fast endpoint
