--- src/PlaceApi.Application/Authentication/Register/RegisterResult.cs
@@ -1,5 +1,0 @@
-using System;
-
-namespace PlaceApi.Application.Authentication.Register;
-
-public record RegisterResult(string UserId, string Email);
