+++ src/PlaceApi.Application/UseCases/Authentication/ForgotPassword/ForgotPasswordCommand.cs
@@ -1,0 +1,6 @@
+using MediatR;
+using Microsoft.AspNetCore.Http.HttpResults;
+
+namespace PlaceApi.Application.UseCases.Authentication.ForgotPassword;
+
+public sealed record ForgotPasswordCommand(string Email) : IRequest<NoContent>;
