+++ src/PlaceApi.Application/UseCases/Authentication/Register/RegisterResult.cs
@@ -1,0 +1,3 @@
+namespace PlaceApi.Application.UseCases.Authentication.Register;
+
+public record RegisterResult(string UserId, string Email);
