+++ src/PlaceApi.Application/UseCases/Authentication/SendConfirmationEmail/SendConfirmationEmail.cs
@@ -1,0 +1,7 @@
+using MediatR;
+using PlaceApi.Domain.Authentication.Entities;
+
+namespace PlaceApi.Application.UseCases.Authentication.Notifications.Confirmation;
+
+public record SendConfirmationEmail(ApplicationUser User, string Email, bool IsChange = false)
+    : INotification;
