--- src/PlaceApi.Web/Endpoints/Authentication/ForgotPassword/ForgotPasswordRequest.cs
@@ -1,6 +1,0 @@
-namespace PlaceApi.Web.Endpoints.Authentication.ForgotPassword;
-
-public sealed class ForgotPasswordRequest
-{
-    public required string Email { get; init; } = null!;
-}
