--- src/PlaceApi.Web/Endpoints/Authentication/ResendConfirmationEmail/ResendConfirmationEmailRequest.cs
@@ -1,6 +1,0 @@
-namespace PlaceApi.Web.Endpoints.Authentication.ResendConfirmationEmail;
-
-public class ResendConfirmationEmailRequest
-{
-    public required string Email { get; init; }
-}
