--- src/PlaceApi.Web/Endpoints/Extensions.cs
@@ -1,45 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.HttpResults;
-using Microsoft.AspNetCore.Identity;
-
-namespace PlaceApi.Web.Endpoints;
-
-public class ValidationsExtensions
-{
-    public static ValidationProblem CreateValidationProblem(IdentityResult result)
-    {
-        Debug.Assert(!result.Succeeded);
-        Dictionary<string, string[]> errorDictionary = new(1);
-
-        foreach (IdentityError error in result.Errors)
-        {
-            string[] newDescriptions;
-
-            if (errorDictionary.TryGetValue(error.Code, out string[]? descriptions))
-            {
-                newDescriptions = new string[descriptions.Length + 1];
-                Array.Copy(descriptions, newDescriptions, descriptions.Length);
-                newDescriptions[descriptions.Length] = error.Description;
-            }
-            else
-            {
-                newDescriptions = [error.Description];
-            }
-
-            errorDictionary[error.Code] = newDescriptions;
-        }
-
-        return TypedResults.ValidationProblem(errorDictionary);
-    }
-
-    public static ValidationProblem CreateValidationProblem(
-        string errorCode,
-        string errorDescription
-    ) =>
-        TypedResults.ValidationProblem(
-            new Dictionary<string, string[]> { { errorCode, [errorDescription] } }
-        );
-}
