+++ src/PlaceApi.Web/Endpoints/V1/Authentication/ForgotPassword/ForgotPasswordRequest.cs
@@ -1,0 +1,6 @@
+namespace PlaceApi.Web.Endpoints.V1.Authentication.ForgotPassword;
+
+public sealed class ForgotPasswordRequest
+{
+    public required string Email { get; init; } = null!;
+}
