+++ src/PlaceApi.Web/Endpoints/V1/Authentication/ResendConfirmationEmail/ResendConfirmationEmailRequest.cs
@@ -1,0 +1,9 @@
+using System.Text.Json.Serialization;
+
+namespace PlaceApi.Web.Endpoints.V1.Authentication.ResendConfirmationEmail;
+
+public sealed record ResendConfirmationEmailRequest
+{
+    [JsonPropertyName("email")]
+    public required string Email { get; init; }
+}
